Add functionality to recalculate alignment zones
In this commit, we've added a new command `recalc_zones` to our CLI tool. This new command recalculates the alignment zones of the font. It is added to facilitate the users to recalculate alignment zones conveniently. Also, exception handling mechanism has been improved by adding `finally` block to ensure the font file closes in all scenarios.