Skip file rename if no change
Added a check in the file rename function in ftcli_utils. This checks whether the output file name is the same as the current file name. If it's the same, it skips the renaming process.

This, in case the command is run twice, avoids renaming font files by adding the #1 prefix